move GetIPCAddress back to pkg/config
